Extract separate extension into Introspection that we then include
